+++ persistence-spring-mongo/src/main/java/io/naryo/infrastructure/store/event/persistence/document/contract/parameter/IntParameterDocument.java
@@ -1,5 +1,7 @@
 package io.naryo.infrastructure.store.event.persistence.document.contract.parameter;
 
+import java.math.BigInteger;
+
 import io.naryo.domain.common.ParameterType;
 import io.naryo.domain.event.contract.parameter.IntParameter;
 import lombok.NoArgsConstructor;
@@ -8,8 +10,9 @@ import org.springframework.data.annotation.TypeAlias;
 @NoArgsConstructor
 @TypeAlias("int")
 public final class IntParameterDocument
-        extends ContractEventParameterDocument<Integer, IntParameter> {
-    private IntParameterDocument(ParameterType type, boolean indexed, int position, Integer value) {
+        extends ContractEventParameterDocument<BigInteger, IntParameter> {
+    private IntParameterDocument(
+            ParameterType type, boolean indexed, int position, BigInteger value) {
         super(type, indexed, position, value);
     }
